Added 2SatAutoDatasetConstruction.py
This file automatically creates two .csv of all the available lakes. To use it, download all the .nc files from the onedrive and place them into a folder called All_Lakes. This will create 2 csv files, one for the X data and one for the Y data. QuickDatasetConstruction.py can then be used to get these ready for use.