Added QuickDatasetConstruction.py
This .py file takes the datasets created by 2SatAutoDatasetConstruction and can remove a subset of it for actual use. (Doing it this way is much faster as the initial dataset can take a few minutes to create) Put the lakes you want to include into lakelist and set filename to be the name the new dataset will be saved as.